Moving out s3 code
