+++ evaluation-tasks/util/s3.py
@@ -2,11 +2,98 @@
 Utility functions for interacting with an S3 bucket.
 """
 
+import os
+import subprocess
+
 import boto3
 from luigi.contrib.s3 import S3Client
 from botocore.client import ClientError
+import luigi
 
 from .luigi import WorkTask
+
+
+class CacheTarCorpus(WorkTask):
+    """
+    If the tar file is cached in S3, we simply retrieve it and untar
+    it, and skip the pipeline.
+
+    If the tar file is NOT in S3, we run the pipeline, create the
+    tar-file, and upload it to S3.
+    """
+
+    task_name = luigi.Parameter()
+    bucket = luigi.Parameter()
+    region = luigi.Parameter()
+    next_task = luigi.Parameter()
+    # TODO: How do we support params for next_task??
+
+    def requires(self):
+        return EnsureBucket(self.bucket, self.region)
+
+    @property
+    def name(self):
+        return type(self).__name__
+
+    def run(self):
+        tarfile = f"{self.task_name}.tar.gz"
+        pathtarfile = f"{self.workdir}/{tarfile}"
+        client = S3Client()
+        s3cache = os.path.join(f"s3://{self.bucket}/", tarfile)
+        if client.exists(s3cache):
+            client.get(s3cache, pathtarfile)
+        else:
+            # If you yield FinalizeCorpus, this task is suspended
+            # and FinalizeCorpus is run, and a LocalFileTarget is
+            # returned.
+            finalize_corpus = yield self.next_task()
+
+            # Tar the file
+            devnull = open(os.devnull, "w")
+            ret = subprocess.call(
+                [
+                    "tar",
+                    "zcvf",
+                    pathtarfile,
+                    # Unfortunately, we have to hardcode
+                    # FinalizeCorpus.workdir()
+                    # because we don't have a Task
+                    # just a Target in finalize_corpus.
+                    self.task_name,
+                ],
+                stdout=devnull,
+                stderr=devnull,
+            )
+            # Make sure the return code is 0 and the command was successful.
+            assert ret == 0
+
+            # Cache to S3
+            print("Putting file to S3")
+            client.put_multipart(pathtarfile, s3cache)
+
+        with self.output().open("w") as outfile:
+            pass
+
+
+class EnsureBucket(WorkTask):
+    """
+    Ensure the S3 bucket exists and is readable.
+
+    This S3 code is pretty gnarly, but I'm not sure it can be made
+    any cleaner.
+    """
+
+    bucket = luigi.Parameter()
+    region = luigi.Parameter()
+
+    @property
+    def name(self):
+        return type(self).__name__
+
+    def run(self):
+        check_bucket(self.bucket, self.region)
+        with self.output().open("w") as outfile:
+            pass
 
 
 def can_access_bucket(s3: boto3.session.Session, bucket: str) -> str:
